requirements.txt
langchain
langchain-core
langchain_community
langchain-openai
python-dotenv
faiss-cpu
pymupdf
